+++ app/src/main/java/com/alorma/diary/ui/activity/AddDiaryEntryActivity.java
@@ -11,7 +11,7 @@ import com.alorma.diary.ui.fragment.AddDiaryEntryFragment;
 
 public class AddDiaryEntryActivity extends BaseBackActivity {
 
-  public static Intent createIntent(Context context, int id) {
+  public static Intent createIntent(Context context, long id) {
     Intent intent = new Intent(context, AddDiaryEntryActivity.class);
     intent.putExtra(Extras.DIARY_ID, id);
     return intent;
@@ -27,7 +27,7 @@ public class AddDiaryEntryActivity extends BaseBackActivity {
       throw new RuntimeException("Should never happen: " + DiaryDetailActivity.class.getCanonicalName());
     }
 
-    int diaryId = getIntent().getExtras().getInt(Extras.DIARY_ID);
+    long diaryId = getIntent().getExtras().getLong(Extras.DIARY_ID);
     AddDiaryEntryFragment fragment = AddDiaryEntryFragment.newInstance(diaryId);
     getSupportFragmentManager().beginTransaction().replace(R.id.container, fragment).commit();
   }
